# Add doc-values aggregator for std deviation aggregation.