## Summary of the changes / Why this improves CrateDB

It improves the performance of the global standard deviation aggregations.
See https://github.com/crate/crate/pull/10048 for more information about the approach.

Without doc-values aggregators:

```
## Running Query:
   Name: global stddev on long
   Statement:
     select stddev(duration) from uservisits
   Concurrency: 1
   Iterations: 500
100%| 500/500 [00:21<00:00, 22.84 requests/s]
Runtime (in ms):
    mean:    42.137 ± 1.729
    min/max: 35.082 → 440.874
Percentile:
    50:   39.567 ± 19.721 (stdev)
    95:   52.029
    99.9: 440.874

## Running Query:
   Name: global stddev on float
   Statement:
     select stddev("adRevenue") from uservisits
   Concurrency: 1
   Iterations: 500
100%| 500/500 [00:23<00:00, 21.51 requests/s]
Runtime (in ms):
    mean:    45.046 ± 0.685
    min/max: 39.185 → 128.267
Percentile:
    50:   43.009 ± 7.816 (stdev)
    95:   57.158
    99.9: 128.267
```

With doc-values aggregators:

```
## Running Query:
   Name: global stddev on long
   Statement:
     select stddev(duration) from uservisits
   Concurrency: 1
   Iterations: 500
100%| 500/500 [00:09<00:00, 53.52 requests/s]
Runtime (in ms):
    mean:    17.106 ± 2.308
    min/max: 11.932 → 547.684
Percentile:
    50:   14.424 ± 26.333 (stdev)
    95:   20.109
    99.9: 547.684

## Running Query:
   Name: global stddev on float
   Statement:
     select stddev("adRevenue") from uservisits
   Concurrency: 1
   Iterations: 500
100%| 500/500 [00:09<00:00, 51.24 requests/s]
Runtime (in ms):
    mean:    18.157 ± 1.545
    min/max: 13.716 → 286.196
Percentile:
    50:   15.705 ± 17.631 (stdev)
    95:   22.723
    99.9: 286.196
```

## Checklist

 - [ ] Added an entry in `CHANGES.txt` for user facing changes
 - [ ] Updated documentation & `sql_features` table for user facing changes
 - [x] Touched code is covered by tests
 - [x] [CLA](https://crate.io/community/contribute/cla/) is signed
 - [x] This does not contain breaking changes, or if it does:
    - It is released within a major release
    - It is recorded in ``CHANGES.txt``
    - It was marked as deprecated in an earlier release if possible
    - You've thought about the consequences and other components are adapted
      (E.g. AdminUI)