@kovrus 2020-08-12
> Should we maybe also add some assertHasDocValueAggregator(... types) check? (Could also be a follow up)

I'll follow up in one of the next doc values aggregators implementations.